[racl_ctrl,dv] Do a minimal change to template racl_ctrl DV
The normal design has a fusesoc core called "MY_BLOCK_sim" which
contains tb.sv and depends on the RTL, and the test and SVA modules.

Here, we do something similar but template the "corefile part" of the
core. Rather confusingly, the guts of this core is called
"racl_ctrl_sim_core.core" (sorry) and they don't need to be templated.
Now the only thing that *does* need to be templated is the thing that
depends on the relevant instantiation.

Note that the templated file (racl_ctrl_sim.core) needs to depend
explicitly on the correct top_racl_pkg. If you don't do this, you
might be unlucky and get the wrong top-level's package pulled in by
some dependency. This is VERY annoying to debug, and should hopefully
be a bit less painful when we switch to upstream fusesoc.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>