[opentitanlib] Respect TPM burstCount
When transmitting large TPM commands, the AP is supposed to inspect the
"burstCount" field of the TPM STS register, in order to determine how
many bytes the TPM can accept in a single SPI/I2C transfer.

Our code was first hard-coded to 32 bytes, later changed to 64 bytes.
In order to be able to handle various TPM implementation, this patch
makes the code use whatever is advertised in burstCount.  (It does not
re-inspect the STS register between each block of bytes, assuming
instead that the TPM has a fixed-size buffer.)

Change-Id: Ib2e7eaf7ef93438ec10be9351567cc35d11dd480
Signed-off-by: Jes B. Klinke <jbk@chromium.org>