[manuf] Initialize boot data cfg OTP during perso.
Enabling scrambling and ECC for the boot data info pages during
individualization was causing perso bootstrap to fail.

This is because ROM attempts to read the boot data before it has been
initialized, causing integrity errors.

The solution is to skip boot data during individualization and apply it
during perso, which is when the boot data is written for the first time.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>
(cherry picked from commit 9eea8d8104f7657044b44a072251902abeb4f827)