[mdbook,doxygen] Fix links to DIF
The `difgen` library was previously called with absolute paths
to the DIF headers. Now that doxygen runs in bazel, all the paths
have been made relative to the root so that the build is deterministic.
However, this means that the difgen also needs to search to DIF header
using relative paths now.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>