[dv, alert_monitor] moving 'active_ping' to alert cfg obj
The flag has been moved from alert_monitor to the agent config object. And will
be used in a later commit.

Reason being:
In some situations the verification environment may need to know if the alert
monitor is sampling a ping.
For example: if there is a random ping before an alert, it can happen the model
expects an alert "too soon" but it may take a few cycles longer due to the ping.

This is typical in scenarios where the stimulus causes an alert intentionally,
but at the time of that caused alert a ping was already on-flight.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>