[clkmgr] Update clkmgr_regwen_vseq to accommodate DJ clocks
The clocks of the TL-UL interface and the monitored clock must _both_
be sufficiently faster than the AON clock for the testing of the
`meas_ctrl` regwen to avoid generating spurious measurement alterts.

The DV environment chooses randomized clock frequencies that are a
few dozens of MHz but the Darjeeling AON clock (reference clock in
the measurement operation) is itself 62.5MHz so it is more than capable
of completing a measurement cycle during the time that it is enabled.

To avoid the successful completion of a measurement period we have
two options - (i) adjust the clock frequencies, or (ii) constrain
the programmed 'lo' and 'hi' thresholds to ensure that all
measurements are valid.

Constraining the thresholds would mean an increased risk of failing to
spot a write that should have been prevented by the `regwen` state,
and does not really solve the problem for Darjeeling because its
narrow counters are liable to overflow.

Thus we choose to ensure that the TL-UL clock is sufficiently high and
disable randomized TL-UL delays for the duration of the test.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>