[pentest] Add ASM trigger high/low
Until now, we've used the pentest_set_trigger_high/low functions.
However, this is problematic as:
- These functions are calling multiple other sub-functions which
  contain a lot of code. Hence, executing this function takes
  a lot of time.
- This is bad for the pentest_set_trigger_high as, once the
  trigger raises, the function still executes for a couple
  of cycles. In these cycles we assume that we are already
  executing our FI target code sequence but instead we
  still execute the trigger function. Hence, we could
  land up attacking the trigger function.
- Similar, for the pentest_set_trigger_low, it takes
  actually quite some time until the trigger falls down.
- As these functions are function calls, they mess around
  with the registers. Hence, we need to initialize the
  registers after the trigger high function. Again, with
  this we could land up attacking the register init.

As measuring this cycle delay is quite challenging, we've
decided to implement an optimized trigger function that
purely consists of ASM instructions. Hence, we also
can move the register initialization before the trigger
as we know which registers are getting touched by the
trigger set macro.

Use with caution as it is not guaranteed that this will
work in future OpenTitan versions.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
Co-authored-by: Siemen Dhooghe <sdhooghe@google.com>
(cherry picked from commit 9b07fac)