[provisioning] inject WAS from orchestrator in benchtop flow
This updates the orchestrator to inject the same WAS during CP and FT to
the rust-based test harnesses. For now we use a WAS of all zeros for
benchtop provisioning flows which are performed in a secure environment.

Signed-off-by: Tim Trippel <ttrippel@google.com>
(cherry picked from commit b14c6b06908748b7febdd0cce4d95aee9db7a69d)