[sw/cryptolib] Add new ibex driver functions for randomness
This commit adds an ibex driver with three functions. Two of the
functions return the value of the RND_STATUS and RND_DATA CSRs.
The third function waits for the RND_STATUS reg to show that the
RND_DATA is valid.

Signed-off-by: h-filali <h.filali98@gmail.com>
(cherry picked from commit 272593adf386e1b5e931a9904dec82674e2216b6)