[rom_ctrl,dv] Tidy up CompareCtrlFlowConsistency check
This wasn't all that clearly thought out, so I've added some detailed
comments about what we're doing (and why). With the existing code,
it *might* have started the checker a second time after it was done.

That presumably didn't happen very often (because the test would have
failed) but we don't really care about that case: the attacker is just
forcing the hardware to confirm what it already told us!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>