[dv, cip_base_scoreboard] Increase alert timeout to 4-cycles
With the changes to the alert monitor there's now an extra 2 cycles delay when
sampling TX/RX alert signals. This has flagged up in some failures which expect
the alert to have happened already.

This fix is just to give 4 extra clock to allow the monitor to send the
seq_item to the scoreboard to carry out the check

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>