[alert,dv] Avoid using state_q and PingSt from alert_esc_if directly
Doing so causes Xcelium to generate an (initially mystifying) error message:

  xmelab: *E,DISVIF (...._vseq.sv,111|48):
    Disallowed virtual interface hierarchical reference object type at 'PingSt'.

This took me a while to understand, but the point is that the vseq was
using e.g. alert_vif.PingSt. What is the type of that expression?
Well, it's state_e, but the enum is defined inside the interface (and
that enum isn't technically part of the "port list" of the interface).

At least, that's what I *think* is going on. There's a discussion
of (I think) the same question at

    https://stackoverflow.com/questions/54627223/

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>