[opentitanlib] add GPIO indicator pin to SpiConsoleDevice
Currently the SpiConsoleDevice polls the DUT's SPI device buffer for
data that it wants to transmit out of the device to the host. To support
less noisy, non-polling environments, this adds an indicator GPIO pin
that can alternatively be sampled to indicate when data is ready to be
transmitted out of the device.

Signed-off-by: Tim Trippel <ttrippel@google.com>