[spi_device,rtl] Move definition of ValidCmdConfig_A
Some tools (Verissimo, at least) generate a lint error because we use
the MainAddress symbol before it's defined as part of the main_st_e
enum.

Move it to come after that and (since I'm touching the lines anyway)
tidy up the indentation to make the structure of the assertion a bit
clearer.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>