[ownership] add ECDSA verification of Activate and Unlock blobs
When debugging new signing flows it could be useful to validate
signature of Activate and Unlock blobs in addition to the Owner blob.

The Activate and Unlock blobs don't have easily identifiable
structure, and even though they are 212 bytes in size at this time,
this could change.

To provide maximum flexibility this patch considers all input blobs of
size different than 2048 possible (TBS, Raw ECDSA Signature)
concatenation and tries to verify the TBS using the provided public
key.

If the input is not Owner blob and the key is not provided, an error
is reported.

ECDSA is still the only accepted signature algorithm at this time.

TEST=successfully verified signed Owner blob with and without a key
     file included in the command line. Successfully verified Activate
     and Unlock blobs using the appropriate keys.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>