[rom_ext] Integrate ISFB into the boot flow
The Integrator Specific Firmware Binding (ISFB) feature allows an owner to
configure an INFO page to hold application firmware binding information.
The binding information allows owners to implement an anti-rollback
policy with a strike-out mask mechanism and per-product binding based
on product identifier words stored in the INFO page.

The owner can then apply an extension to the application firmware manifest
to specify the strikeout binding (ie: number of unstruck words) and a
series of product expressions which match against the product words.

This feature is currently documented in this
[mini-spec](https://docs.google.com/document/d/19oQtqgDdFU6qunbwgkwNaizYUNCzdLVqmnBpLaDvX5I/edit?pli=1&tab=t.0#heading=h.bw1472fp996n).

Signed-off-by: Miguel Osorio <miguelosorio@google.com>