[sw/crypto] Add new struct for word buffer with byte length field
This commit adds a constand and non constant version of a struct
that contains two fields. Data and len. Data is a 32 bit size word
buffer and len contains the buffer length in bytes.

These two structs are added to their counterparts:
- 32 bit size word buffer and buffer length in words
- 8 bit size word buffer and buffer length in bytes

The new struct was created because just using the 32 bit word size
struct could be cause for confusion due to namings and comments.

Signed-off-by: h-filali <h.filali98@gmail.com>