[sw/crypto] Use word buffers for input and output to the hardened memcpy
This commit changes the input and output buffers of the hardened memcpy
function to 32 bit word size buffers with the buffer size expressed as
its byte length.

This is done to avoid running into issues where the bytes are misaligned
with respect to 32 bit word boundaries. This allows for a simpler hardened
implementation of memcpy. The buffer length still needs to be expressed in
bytes however such that the padding is correct.

Signed-off-by: h-filali <h.filali98@gmail.com>