[darjeeling] Fix SRAM Ctrl scrambled access sequence
The synchronization method between the DV vseq and the
SW test has changed; migrate the changes from the
modified Earl Grey implementation.

Update the sampling intervals for Darjeeling clock
frequencies because the 1ns delay used on Earl Grey
will no longer be reliable. Also reduce the burden of
the periodic sampling for those clock frequencies that
are known to be lower.

Modify the `req` signal indexes to be appropriate for
the Darjeeling OTP control.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>