[opentitanlib] add SPI console frame reset function
This adds a function to reset the SPI console frame counter to enable
use of the TX indicator GPIO pin with test firmware that issues several
resets.

When the TX indicator GPIO pin is used in conjunction with the SPI
console, the bootup synchronization message exchange is not used.
Instead the host expects the device to use buffer address 0 as the
initial console frame offset. If this is not the case the host will get
wedged trying to read an invalid frame. Additionally, the host keeps
track of the last read frame so it knows where to find the next frame.
When the chip resets, the host needs to reset the frame counter if it is
using the TX indicator pin feature of the SPI console to start reading
from address 0 as well.

Signed-off-by: Tim Trippel <ttrippel@google.com>
(cherry picked from commit 810be6948d832e647f174acfb2c2e3f2d15d8d2b)