[rom_ctrl,fpv] Define a helper property for StabilityChkkeymgr_A
The target property was not converging. You can generate a 3-cycle SST
example with

    prove -property tb.dut.StabilityChkkeymgr_A -sst 3

and the result shows something where `in_state_done` is true in the
FSM, but `checker_good` changes. The root problem seems to be that the
k-induction search doesn't know that the compare module thinks it is
done when the surrounding checker FSM does.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>