[rom_ext, ownership] Adjust owenr block signature verificaiton
Ownership with PQ keys require that ownership signatures are detached
from the owner block and be stored in flash (PQ signatures are too
big to fit into the 2K owner block struct).

Reading from flash correctly requires that flash be configured and flash
configuration is stored in the owner block.  In order to accept an
update or perform a transfer, we have to initialize the flash according
to the current owner configuration _before_ attempting to find the
signature in flash and verify it.

Adjust `{un,}locked_init` to check the validity of page 0 and then
initialize the flash config before checking for updates in page 1.

Signed-off-by: Chris Frantz <cfrantz@google.com>