[rv_plic, fpv] Rsp and Data intg undetectable in reg_if
rv_plic_reg_top instantiates tlul_adapter_reg with EnableDataIntgGen
and EnableRspIntgGen as 0. So, rsp_intg and data_intg is always zero
and jasper recognize that undetectable.

Tweak the instantiation of tlul_adapter_reg in a way applied in the
PR will allow jasper to not look at the piece of unneccessary code
and mark it as undetectable.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>