[rom_ctrl,rtl] Fix and strengthen the TlAccessChk_A assertion
The statement wasn't actually true (because the pwrmgr_data_o.done
signal isn't one-way), but you can tweak it to make things true.

Note that one property that *is* true (but not proved) is that we'll
never see a response with d_valid && !d_error until after
pwrmgr_data_o.done has become MuBi4True (even if it changes back again
afterwards).

That seems a bit fiddly for an embedded statement in some RTL, so I'm
not proving it here.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>