[manuf] remove unused SKU config field
As part of the E2E provisioning flow, the FT test harness can test for
the presents of a success message printed over the console by the
transport image (i.e., owner firmware) that is loaded during the
provisioning process. This string is configured in the orchestrator.py
HJSON configuration file, and was duplicated in the SKU configuration
dictionary. This duplicate was unused, and hence is removed by this
commit.

Signed-off-by: Tim Trippel <ttrippel@google.com>
(cherry picked from commit 058f4c4b2fb0b0817cb7d943e60b554d4f395ab9)