[dv, rv_timer] Add min/max register values test to hit FCOV corners
The max test will set the step, prescaler and comparison value to be the maximum
value possible. The timer will still be set at a gap up to 5 units from the
maximum just so interrupts can be triggered both by incrementing the counter but
also by straight writing a value which makes mtime >= mtimecmp.

The min test will set the prescaler and the timer to 0. The timecmp register
will be set to a value of up to 5, just so the interrupt is not always produced
directly from the configuration.
The step is randomised to be either 0 or 1. This is to allow the counter to
increase half of the time since the counter increase by: count = count + step.

Signed-off-by: Antonio <amz@lowrisc.org>