[dv, rv_timer] Reduce the clock cycles to count by 1000 for random_vseq
This will allow to quicken up our simulations rather than having the counter
increase up to 5 million ticks.
For the purpose of the verification, we are mostly interested in contrived
scenarios within the spec on boundary conditions.
Plus, it seems unlikely any bug will be found by just forcing the counter to
count through a larger interval.

Signed-off-by: Antonio <amz@lowrisc.org>