[individualize] add AST config version flash info field
Previously, the AST config version field was provided by the host during
the FT individualization step. However, since the AST values are
programmed to flash info page 0 during the CP stage, the AST
configuration version is also known then. Therefor we add a field to
flash info page 0 to store this field and overwrite this portion of the
FT device ID during individualization before the device ID is written to
OTP.

Signed-off-by: Tim Trippel <ttrippel@google.com>