[sw/cryptolib] Remove out dated TODOs about using hardened memcpy
After discussion the decision for the TODOs removed in this commit
is to remove them. We want to support byte aligned copying of data
for the use cases of AES messages and drbg seed material. The hardened
memcpy only accepts word aligned data which we do not want to
guarantee here.

Signed-off-by: h-filali <h.filali98@gmail.com>