[dv,ate] add ATE test to perform bootstrap erase
This adds a simple ATE DV test that puts the DUT in bootstrap mode and
then drives the correct SPI sequence to erase all of flash, which is the
first part of performing a bootstrap operation. This will aid ATE
bringup work.

This is a cherry-picked version of commit b59973a6e1, which was added
to the earlgrey_1.0.0 branch in May 2025. The cherry-pick has been
edited somewhat to avoid having to repeat code in the base vseq.

Co-authored-by: Tim Trippel <ttrippel@google.com>
Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>