[util] add script to compute committer stats
This adds a script to compute committer stats for the OpenTitan
repository using the GitHub GraphQL API. Currently this script can only
compute stats for the last 100 pages of commits, but it is a good
starting point to iterate on to create a final product that can compute
stats across the whole history of the repo.

In the future, it may be convenient to additionally filter the results
to only aggregate stats for contributors that are no officially
committers yet, as this script will be used to automate committer
proposals.

Signed-off-by: Tim Trippel <ttrippel@google.com>