[util] Add optional args to the committer stats script
This commit adds two optional arguments to the commiter stats script.
One is a `-l`/`--log-progress` flag which sets a global flag to enable
debug logging for every paginated query, since queries over a large repo
like OpenTitan can take a fair while to compute. This lets users opt in
to more visible progress.

Also adds a `-p`/`--pages` option that can be used to limit the number
of pages used in each query, since this functionality was essentially
already available and just needed exposing. This can be used for testing
purposes, to run only on a subset of the repository's history for
quicker prototyping.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>