[util] Rate limit & error handling in committer stats script
This commit adds better rate-limit and error handling functionality to
the committer stats script. If a rate limit response is received we
sleep until the requested time to respect the rate limit, and log if
logging is enabled. We also check the HTTP status code and JSON
response for HTTP / GraphQL errors to better handle these cases.

Also, if we get a connection timeout from the Python requests library due to
inability to reach the GitHub API, retry with exponentially increasing
timeouts between posting the request and waiting.

We have a silently retry with 1 second to account for poorer connections.
After that, we exponentially increase the time we wait until the response is
received or the user cancels the script.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>