[util] Add option to filter existing committers from commiter stats
This commit adds a command-line option that was suggested to filter out
those who are already in the committers file from the generated list of
commiter stats in the report, so that it can be used to identify
potential candidates to become committers.

This is limited however in that it only looks for current contributors.
In the future it would be nice to extend this, either by giving it a
full historic list of committers or by adding a feature to search the
git history, such that we can also exclude people who were previously
committers but are not any more.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>