[util] Find historical committers for committer stats
This commit extends the existing logic in the committer stats script
that filters out committers in the COMMITTER file from the results, to
find a complete list of all committers in the (tracked) history of the
OpenTitan project by searching local git revision of the COMMITTERS file
and parsing the contents at each point.

This is more useful, as it means that using the `--filter` option will
no longer return several historic committers.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>