[dv, i2c] Remoce explicit `ref` keyword from uvm_events in function argument
uvm_events are already a reference so the `ref` keyword is not needed. Removing
`ref` removes the Xcelium warning (FJNRFAW)

According to the LRM (section 9.3.2) it's illegal to refer to formal arguments
passed by reference. Based on the wording in the LRM, it seems this refers
explicitly for when the argument to a task has the `ref` keyword, as there is no
mention for a class pointer which is indeed a reference.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>