[tl,dv] Remove pre_body specialisation in tl_device_seq
It's rather simpler to put the "thing to do before body" into... the
start of the body task!

The pre_body task also triggers warnings from lint tools because it
doesn't always get called. (It doesn't get called when the sequence is
started with `uvm_do: see the notes in the documentation for
uvm_sequence_base)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>