[tl,dv] Remove pre_body specialisation in tl_device_seq
I don't think this really made much sense. If I understand correctly,
the idea was to clear the "stop" flag just before the body() task
started.

The only way that the stop flag could be set is by calling the
seq_stop task. Amusingly, there was a race condition from the
pre_start behaviour, which I worked around with 18d882d. But it's even
easier not to do the silly thing at all. So don't!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>